Implemented ETag and max cost criteria

+++ src/main/java/in/co/trish/marketscan/persistence/repositories/specifications/ProductSpecification.java
@@ -16,30 +16,30 @@ import in.co.trish.marketscan.persistence.entities.Product;
 import in.co.trish.marketscan.web.representation.read.ProductSearchCriteria;
 
 public class ProductSpecification implements Specification<Product> {
-
+	
 	private ProductSearchCriteria criteria;
-
+	
 	public ProductSpecification(ProductSearchCriteria criteria) {
 		this.criteria = criteria;
 	}
-
+	
 	@Override
 	public Predicate toPredicate(Root<Product> root, CriteriaQuery<?> query, CriteriaBuilder builder) {
-
+		
 		List<Predicate> predicates = new ArrayList<>();
 		
 		if (!StringUtils.isEmpty(criteria.getBrand())) {
-			predicates.add(builder.equal(root.<Brand>get("brand").<String>get("name"), criteria.getBrand()));
+			predicates.add(builder.equal(root.<Brand> get("brand").<String> get("name"), criteria.getBrand()));
 		}
 		if (!StringUtils.isEmpty(criteria.getName())) {
-			predicates.add(builder.like(builder.upper(root.<String>get("name")), "%" + criteria.getName().toUpperCase() + "%"));
+			predicates.add(builder.like(builder.upper(root.<String> get("name")), "%" + criteria.getName().toUpperCase() + "%"));
 		}
-		if(criteria.getMax_cost() > 0){
-			predicates.add(builder.lessThanOrEqualTo(root.join("offers").<Long>get("price"), criteria.getMax_cost()));
+		if (criteria.getMax_cost() > 0) {
+			predicates.add(builder.lessThanOrEqualTo(root.join("offers").<Double>get("price"), criteria.getMax_cost()));
 		}
 		
 		query.distinct(true);
-		return builder.and(predicates.toArray(new Predicate[]{}));
+		return builder.and(predicates.toArray(new Predicate[] {}));
 	}
-
+	
 }
